feat: first pass at a unique Minecraft stack
